backend/requirements.txt
fastapi
uvicorn
python-multipart
httpx
python-pptx
pydantic
